+++ lib/NodeRpcProxy/NodeRpcProxy.cpp
@@ -23,8 +23,8 @@
 #include <system_error>
 #include <thread>
 
-#include <HTTP/HttpRequest.h>
-#include <HTTP/HttpResponse.h>
+#include <Http/HttpRequest.h>
+#include <Http/HttpResponse.h>
 #include <System/ContextGroup.h>
 #include <System/Dispatcher.h>
 #include <System/Event.h>
