+++ lib/Rpc/HttpServer.cpp
@@ -21,7 +21,7 @@
 #include "HttpServer.h"
 #include <boost/scope_exit.hpp>
 
-#include <HTTP/HttpParser.h>
+#include <Http/HttpParser.h>
 #include <System/InterruptedException.h>
 #include <System/TcpStream.h>
 #include <System/Ipv4Address.h>
